Change "Stage" folder name to lowercase
The current "Stage' folder name causes issues while trying to use it on a sensitive case operative system, like linux, due to "Stage" being lowercase on the .ini, file.

instead of changing all texture entries in that folder, we all just need to change the folder name itself to lowercase so it matches with the .ini.

This doesn't affect windows.